+++ src/ulog.c
@@ -9,16 +9,22 @@
 #include "uattrs.h"
 #include "ucolor.h"
 #include "udebug.h"
+#include "ulib_ret.h"
 #include "ulib_ret_t.h"
+#include "ulock.h"
+#include "ulog_p.h"
 #include "umetrics.h"
 #include "unumber.h"
 #include "ustream.h"
 #include "ustring.h"
 #include "utime.h"
-#include "uutils.h"
 #include <stdarg.h>
 #include <stddef.h>
 #include <string.h>
+
+static ULog main_logger;
+ULog *const ulog_main = &main_logger;
+static ULock shared_lock;
 
 enum { LEVEL_COUNT = 8 };
 static char const *level_str[LEVEL_COUNT] = {
@@ -28,6 +34,17 @@ static char const *level_color[LEVEL_COUNT] = {
     NULL,        UCOLOR_TRACE, UCOLOR_DEBUG, UCOLOR_PERF,
     UCOLOR_INFO, UCOLOR_WARN,  UCOLOR_ERROR, UCOLOR_FATAL,
 };
+
+ulib_ret p_ulog_init(void) {
+    ulib_ret ret = ulock(&shared_lock);
+    if (ulib_is_err(ret)) return ret;
+    *ulog_main = ulog_default();
+    return ULIB_OK;
+}
+
+void p_ulog_deinit(void) {
+    ulock_deinit(&shared_lock);
+}
 
 ULIB_INLINE ulib_ret begin_color(ULog *log, char const *color) {
     return color && log->color ? uostream_write_cstring(log->stream, color, NULL) : ULIB_OK;
@@ -81,6 +98,7 @@ ULog ulog_default(void) {
         .color = ULOG_COLOR,
         .stream = uostream_stderr(),
         .handler = ulog_write_event,
+        .lock = &shared_lock,
     };
 }
 
@@ -173,17 +191,18 @@ ulib_ret ulog_write_newline(ULog *log) {
     return uostream_write_literal(log->stream, "\n", NULL);
 }
 
-ULog *p_ulog_main(void) {
-    static ULog log = ulib_zero_init;
-    if (ulib_unlikely(!log.stream)) log = ulog_default();
-    return &log;
+static ulib_ret handle_event(ULog *log, ULogEvent const *event) {
+    if (!log->lock) return log->handler(log, event);
+    ulib_ret ret = ULIB_OK;
+    ulock_with (log->lock) ret = log->handler(log, event);
+    return ret;
 }
 
 ulib_ret p_ulog(ULog *log, ULogEvent event, char const *fmt, ...) {
     if (!log->handler) return ULIB_OK;
     event.msg.fmt = fmt ? fmt : "";
     va_start(event.msg.args, fmt);
-    ulib_ret ret = log->handler(log, &event);
+    ulib_ret ret = handle_event(log, &event);
     va_end(event.msg.args);
     return ret;
 }
@@ -198,7 +217,7 @@ ulib_ret p_ulog_metrics(ULog *log, ULogEvent event, umetrics_flags flags, char c
     event.data = &data;
     event.msg.fmt = fmt ? fmt : "";
     va_start(event.msg.args, fmt);
-    ulib_ret ret = log->handler(log, &event);
+    ulib_ret ret = handle_event(log, &event);
     va_end(event.msg.args);
     return ret;
 }