+++ src/ulog_p.h
@@ -1,0 +1,21 @@
+/**
+ * @author Ivano Bilenchi
+ *
+ * @copyright Copyright (c) 2026 Ivano Bilenchi <https://ivanobilenchi.com>
+ * @copyright SPDX-License-Identifier: ISC
+ */
+
+#ifndef ULOG_P_H
+#define ULOG_P_H
+
+#include "uattrs.h"
+#include "ulib_ret.h"
+
+ULIB_BEGIN_DECLS
+
+ulib_ret p_ulog_init(void);
+void p_ulog_deinit(void);
+
+ULIB_END_DECLS
+
+#endif // ULOG_P_H